Add files via upload
Each graph holds a different value and measures for the CA Schools. The 1st graphic compares the average math scores of both 6th and 8th graders, from the graph we can see that the 6th graders average is better than 8th. The 2nd graphic compares the total math score of each student in 6th and 8th grade. From the box plot we can see that the 8th graders had some outliers to the graph. The 3rd graphic shows the average student per county. We can see that a certain county has the majority of students, we can use this data for further research if needed.